*mikesamuel/bug-761@2580 | mikesamuel | 2008-10-13 19:31:29 -0800 (Mon, 13 Oct 2008)
Description:

*mikesamuel/bug-761@2428 | mikesamuel | 2008-09-26 08:39:10 -0800 (Fri, 26 Sep 2008)

Description:

Fix bug 761: global scope in disjoint script elements

    <script>var x = 3;</script>
    <script>x = x + 1;</script>
    
    in browsers, all scripts are in the same global scope, so all those x's are
    the same object.
    
    in the testbed with valija off, it complains "Cannot assign to a free
    module variable".  this makes it hard to use onevent= attributes, since in
    this code:
    
    <a onclick="clicky()">click</a>
    <script>function clicky(e) { return true; }</script>
    
    the two clickys are not the same object.
    
    valija doesn't have this problem, so I think the fix is "finish valija".
    
    but maybe there should be a test case for this anyway.


I moved the code that creates a ModuleEnvelope to the ConsolidateCodeStage.
I then reworked the ValijaRewriter so that returnLast() emits its output to a moduleResult___ variable, so that translated code could continue to run after the result.
I then added a TranslatedCode pseudo node (which is consumed by ValijaRewriter) so that returnLast can skip compiled HTML and CSS when trying to determine the last expression in a module.    
Modified the module rule in CajitaRewriter so that "Array" and "Object" are always imported first, and always before the abbreviated [] or {} syntax can be used.

I think this simplifies things a bit since we now no longer have nested modules, and don't have to pull at things around module boundaries.



git-svn-id: https://google-caja.googlecode.com/svn/trunk@2623 360991c0-9f3b-0410-ba87-8b362a00edf4