Issue 11644  http://codereview.appspot.com/11644 Fixes Valija typeof. Adds disabled tests for record inheritance fixes.
Partial fix for bug #814:
Changes Valija's typeof rule so that "typeof [].push" is "function" in Valija.
Disabled tests 953 & 956, which need to be fixed in order to support
non-additive monkey patching. CL 10273 contains a first attempt at
fixing these, which we're not committing since 1) it doesn't work on
Rhino for undiagnised reasons, and 2) it causes a severe performance
loss on the v8 benchmarks. Instead, we here commit but disable the
tests.
Other improvements: Expose isFrozen as cajita.isFrozen. Comment and
whitespace improvements.

R=ihab.awad@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@3155 360991c0-9f3b-0410-ba87-8b362a00edf4