Issue 17068  http://codereview.appspot.com/17068 Issue 938: Setting background-image does not work on IE6/7
http://code.google.com/p/google-caja/issues/detail?id=938
    The problem is the way caja is quoting ':' in css as r'\3A'.
    IE6+7 interprets r'url(http\3A...)' as expected when the string is
    literal text in a page, but it doesn't work when that string is
    assigned to .cssText

    iirc, the r'\3a' quoting was paranoia about weird behavior when a
    css parser tries to do error recovery.  in particular, IE will
    ignore quotes when looking for a place to restart.

    I haven't yet found any type of : quoting that will work for IE.

    note, this is a little awkward to test in the testbed, since all
    urls get rewritten to bounce through a nonexistent proxy.

I changed Escaping to not escape the characters ':' and '='.  This is
sufficient to allow properly formed URLs, since those are the only
characters previously escaped that cannot be converted to %xx form in
a URL without change of behavior.

I don't think this increases the likelihood of code in string problems
in CSS since code in string schemes usually require use of brackets or
quotes, all of which are still escaped.

In CSS, starting a declaration group, using the expression(...)
construct, embedding a url(javascript) and -moz-binding constructs
require brackets or quotes.

R=jasvir@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@3264 360991c0-9f3b-0410-ba87-8b362a00edf4