Issue 8877  http://codereview.appspot.com/8877 Rough draft of design for a fleshed out DOMita.
The package.html describes the code layout for the fleshed out DOMita.
It lists the files and broadly what each will do, talks about how the
tamed node types are derived from the definitions in the DOM Level 2
spec, and covers topics such as parentNode and ownerDocument
navigation in valija and cajita; and focus handling.

R=erights@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@3268 360991c0-9f3b-0410-ba87-8b362a00edf4