Issue 52069  http://codereview.appspot.com/52069 Issue 1018:  Misc issues with JS Html Sanitizer on JScript
RichB reports

I've been playing with the JavaScript HTML sanitizer in Caja on
Windows outside of a browser. Some things I've learnt:

1) It works great on Windows under the Windows Script Host.
2) It works server-side in Classic ASP using the JScript language,
only if you concatenate the <!-- in OUTSIDE_TAG_TOKEN so it doesn't
look like an SGML comment...
3) It compiles in JScript.Net if eqRe is defined as /\=/g instead of /
=/g (!).
4) It throws an exception if you do something like <div
style1='xxx'>blah</div> It seems to barf on the '1'.


Fixes issues reported in (2), (3), and (4).

R=jasvir@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@3458 360991c0-9f3b-0410-ba87-8b362a00edf4