Issue 52088  http://codereview.appspot.com/52088 Moves testbed and cajoling service out of opensocial
This is the first of two CLs to have a clearer separation between caja
and tools that use it (eg opensocial)

Moves the testbed applet and cajoling service out of the opensocial package.

Another CL removes the dependence between the applet and the remaining classes
in opensocial and removes them.

R=ihab.awad@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@3470 360991c0-9f3b-0410-ba87-8b362a00edf4