fix some debugmode issues http://codereview.appspot.com/88132
1. you can't use debugmode twice on the same page.
http://code.google.com/p/google-caja/issues/detail?id=637

cajole this in the testbed with debug on:
  <script>function f(){}</script>
then try to cajole it again.
the second time fails "___ reused with different debug symbols"

this change fixes that by blithely allowing you to reuse ___ with
different debug symbols.  this is not necessarily the right fix,
because two modules coexisting on the same page will stomp on each 
other's debugging symbols, but this simple fix reduces frustration
with the testbed in the common case.

2. event handlers can screw up the debug stack and raise a
spurious exception.
http://code.google.com/p/google-caja/issues/detail?id=828

the current code assumes that event handlers are always
top-level entry points, and it clears the debug stack at
every event, but sometimes an event handler can fire
while we're in middle of executing some other js,
and on return from the eventhandler, debugmode tries to
pop from the empty stack, causing an exception.

this change fixes that in two ways:
a. event handlers will not clear the debug stack.
b. debugstack underflow will not raise an exception.

R=metaweta@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@3566 360991c0-9f3b-0410-ba87-8b362a00edf4