Enhance handling of forms http://codereview.appspot.com/96198
* Fix the following bugs:

  http://code.google.com/p/google-caja/issues/detail?id=1064
  http://code.google.com/p/google-caja/issues/detail?id=1091

* Make the Domita tests use eval() to run code, allowing the eval()-d
source to be seen by the debugger in Firefox. (Using new Function() or
adding a <script> node both cloak the code from the debugger.)

* Add initial support to DomitaTest to report what elements caused
out-of-bounds clicking or waiting to happen.

* In the process, refactored Domita's taming of NodeList and
HTMLCollection objects a bit.

* Refactored the testbed HTML so that it does not put the rendering of
the cajoled stuff in a big <form>. This is because nested <form>s are
not really supported, and so if the cajoled stuff itself contained a
<form> node, things would break.

R=jasvir@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@3657 360991c0-9f3b-0410-ba87-8b362a00edf4