Distinguish valueless attribs (checked, selected) in the HTML attribute schemas http://codereview.appspot.com/111042
HTML allows some valueless attributes.
  <input checked>
is equivalent to
  <input checked="checked">

The HTML schema table distinguishes these by putting the attribute
name as the only allowed value.

This change modifies the perl script that derives the JSON doc, and the
JSON HTML attributes, and it adds programmatic support via the HTML.Attrib
class.

It also changes the value criterion used to check untyped HTML attributes
values to allow the value criteria to be converted to Javascript regular
expressions so that they can be evaluated client side.

R=metaweta@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@3658 360991c0-9f3b-0410-ba87-8b362a00edf4