Modify CSS Validator to coerce invalid font properties to valid ones. http://codereview.appspot.com/115053
http://code.google.com/p/google-caja/issues/detail?q=1105

This partly addresses issue 1105
  Disallowed property values:
    font: x-small;
    font: 100%;
    font: arial;
by coercing these to, respectively,
    font-size: x-small;
    font-size: 100%;
    font-family: arial;

It takes care not to coerce things like
    font: inherit  // Illegal, but has a clear meaning
    font: caption  // Legal

It does coerce some typos, and gives an error message at the
right place, but possibly misleading since
    font: waption
is coerced to
    font-family: "waption"
instead of being reported with a bad property value message.

font:waption is probably treated as an unknown font-name by most
browsers, and previously we would not report a bad value message
for the equally widgy
  font: 12px waption

See the added testcases for details.

R=metaweta@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@3713 360991c0-9f3b-0410-ba87-8b362a00edf4