onclick= set by innerHTML is broken http://codereview.appspot.com/115082
http://code.google.com/p/google-caja/issues/detail?id=588

domita has support for setting very simple onevent= handlers
when assigning to innerHTML.  for example, this is supposed
to work:

  <div id='a'></div>
  <script>
    var a = document.getElementById('a');
    a.innerHTML = '<input type="button" onclick="foo();">';
    function foo () { cajita.log(1); }
  </script>

however, this has been broken since valija was created,
because plugin_dispatchEvent knows how to find a cajita 'foo',
but not a valija 'foo'.

this change teaches plugin_dispatchEvent about valija 'foo'.

R=mikesamuel@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@3722 360991c0-9f3b-0410-ba87-8b362a00edf4