fix 1057, class= attributes are too restrictive http://codereview.appspot.com/89073
1. <div class="!@#$%"></div> is valid html, and it was accepted by
an earlier version of caja, but now it isn't.  This change removes
the constraint.  This is important because some coders and libs
use @class to store metadata about the node.

2. <input name="foo[]"> is not valid html, but it's commonly used,
and browsers allow it.  This change extends the range of allowed
characters in ids, and adds a browser-expectations test to verify
that browsers do not mangle weird ids.

3. browser-side sanitization didn't handle IDREFS attribute values
correctly.  This change fixes that.  At the moment, IDREFS are not
that useful, but they'll become more useful when the various aria
attributes are supported.

4. cajoler-side sanitizer mangles GLOBAL_NAME attributes, but
browser-side sanitizer doesn't.  This change makes them
both mangle.

5. browser-side and cajoler-side policies were inconsistent
about whether <input name='foo__'> is allowed or not.
This change removes the inconsistency.

6. some minor changes to jsunit.js to make it more convenient
on various browsers.

R=mikesamuel@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@3743 360991c0-9f3b-0410-ba87-8b362a00edf4