Change representation of attribute names in client-side HTML schemas. http://codereview.appspot.com/124115
On the server, we use a::href style attribute names in schemas with
two colons between element and attribute names so that we can allow
prefixed names, e.g. html::xml:lang and svg:rect::id.

But on the client, we use one colon, e.g. a:href.

This is a semantics preserving change which brings our client side
representation in sync with the server side one.

R=jasvir@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@3788 360991c0-9f3b-0410-ba87-8b362a00edf4