support window.onload http://codereview.appspot.com/150070
Domita supports window.addEventListener('load')
but doesn't do anything with window.onload.

This adds support for window.onload.
It's not exactly identical to browser handling
of window.onload, because the order of
running onload handlers is different.

Browsers act as if setting window.onload were a
call to removeEventListener then addEventListener, 
and all the onload handlers get run in the order
they were registered.  I didn't see a reason to
add all the code necessary to duplicate that order.

R=mikesamuel@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@3835 360991c0-9f3b-0410-ba87-8b362a00edf4