Fix CSS schema to distinguish top-border from all borders http://codereview.appspot.com/189043
Fixes CSS signature of the border property to reduce the chance of
confusion between a color that applies to all borders and one that
applies only to the top.  This replaces the confusing CSS2.1
specification with the clearer and otherwise clearer version used in
CSS3.

Fixed a nasty non-local side effect bug in SyntheticPropertyMap.
 
Also makes it easier to test CSS color remapping, and makes the hex
encoding lowercase to be consistent with Escaping.java.

R=jasvir@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@3942 360991c0-9f3b-0410-ba87-8b362a00edf4