Fix misplaced directive prologues in rewriter output http://codereview.appspot.com/181060
Some Quasi Rewrite rules of the form
  { @ss* }  ->  { foo(); @ss* }
are problematic because when *ss; matches a directive prologue, the
output will include the prologue in a context where it is likely to be
misinterpreted.
E.g. if @ss* in the above matches "use strict"; bar(), then the output
would be
  { foo(); "use strict"; bar() }
and when rendered and reparsed, the "use strict"; is interpreted as a
string, not a directive prologue.

R=ihab.awad@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@3962 360991c0-9f3b-0410-ba87-8b362a00edf4