Change object ctor quasis http://codereview.appspot.com/207083
Cleans up a number of issues with object constructor quasis that make
it easier to express module metadata.

This is part of the caching pipeline changes.  It makes it easier to rework
modules in CajitaRewriter and ModuleManager so that I can break a page
into separately cajolable modules, so that we can cache the expensive
rewriting stage for common resources like libraries and for portions
of a web application that do not vary from user to user.  It will be
followed by a CL that swaps the order of ConsolidateStage and
ValidateJavascriptStage.

Previously,
   { @x: @y }
matched any number of properties.
This was confusing as indicated by some of the comments around the
uses of Rules.single{Match,Subst} in the rewriters.

Now
  { @x: @y }
matches a single generic property.
  { x: @y }
matches an object with one property named x.
  { @x__: @y, @a*: @b* }
matches an object with at least one property that ends in __ and any
number of other properties.

  * Change QuasiBuilder and QuasiNode implementations so that
    { k: v } matches a single key/value pair, where k and v can be
    any quasi (though y is restricted to one that can match strings.
  * Further change those so that { @x*: @y* } matches any number
    of key/value pairs not matched by the more specific construct
    above.
  * Introduce a new object quasi node that ties the single property
    and multiple quasis together.
  * Got rid of Rule.{match,subst}SingleMap which are rendered moot
    by the new object quasi syntax.
  * Abstract out equivalence testing in SimpleQuasiNode and
    make sure that StringLiterals match by decoded value.
    Previously, the quasi ("foo") would not match ("f\oo").
  * Add a static TestBuildInfo.getInstance method so that
    "TestBuildInfo." does not auto-complete to BuildInfo.getInstance.
  * Added checks for RuleChain so that it buckets object literal
    quasis more finely.
  * Cleaned up code in DefaultValijaRewriter that was making
    StringLiterals the hard way.
  * Changed the quasi parse cache to be thread safe.

R=ihab.awad@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@3995 360991c0-9f3b-0410-ba87-8b362a00edf4