fix the testbed http://codereview.appspot.com/951041
somemtimes I still use the testbed for quick comparison of
behavior of different caja versions.  currently it's broken.

1. testbed html said "valija.co.js" instead of "valija.out.js"

2. testbed.jar was including about 55M of junk (the non-junk
is about 2M).  the junk was causing the testbed to fail to
work for me.  exception in firefox is
   java.lang.NoClassDefFoundError: nu/validator/htmlparser/common/TokenHandler
I'm not sure why the junk was causing this problem,
but removing the junk fixes the problem.

3. htmlparser.jar was twice the size necessary, because
it was including all the .class files twice.

R=mikesamuel@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@4056 360991c0-9f3b-0410-ba87-8b362a00edf4