Message cleanup and config options to make it easier to cajole jquery tests http://codereview.appspot.com/1850042
This changes HtmlAttributeRewriter to fix a malformed URI error that is actually the result of a policy rejection.  I will follow up on this to add options to specify a URL policy as a stopgap measure until we have JS uri policies.

This also adds a config option that allows one to specify the base directory under which file: URIs can be resolved.

R=metaweta@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@4182 360991c0-9f3b-0410-ba87-8b362a00edf4