Fix DomitaTest valija testReadOnlyNotEditable to actually test. http://codereview.appspot.com/1896043
The documentRO variable was not actually visible in DomitaTest in Valija mode, because it was not marked readable on the outers object; therefore testReadOnlyNotEditable was getting its expected failures from documentRO being unbound rather than edits failing.

This change makes the test check that access through documentRO work, and makes domita_test.html create documentRO using cajita.setPub so that it is visible.

R=jasvir@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@4209 360991c0-9f3b-0410-ba87-8b362a00edf4