Preserve XMLNS declarations in the DOM Parser http://codereview.appspot.com/2183044
http://code.google.com/p/google-caja/issues/detail?id=1115

There've been a number of bugs pointing out that we remove XMLNS declarations in the output unless
there's an element that uses them.

So
  <div xmlns:svg="..."></div>
is rendered as
  <div></div>

This doesn't seem an error to me, but this change preserves XMLNS
declarations in the DOM and allows the renderer to output them as long
as they don't mask an existing declaration.

I'm not convinced this is a good idea though.
Some have argued that they are necessary for scripts.  E.g. that
  <div xmlns:foo="bar" id="baz"></div>
will behave differently when
  document.getElementById('baz').innerHTML = '<foo:boo/>';

I don't believe this.  Does anyone have a reference to a standard or an
actual testcase that demonstrates this.

R=jasvir@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@4291 360991c0-9f3b-0410-ba87-8b362a00edf4