improve playground policy tab http://codereview.appspot.com/4545054
1. in the playground policy tab, it's easy to forget to press the 'apply'
button.  the actual policy used for cajole is hidden state, not necessarily
the same as what's shown in the policy tab, and there's no clear indication
of what the hidden state is.

so there are two UI options:

a. expose the hidden state better: disable the 'apply' button, enable it
when the textarea gets changed, add a 'revert' button to revert the
textarea to the hidden state, add a warning if 'cajole' is pressed while
there are unapplied policy changes, etc.

b. get rid of the hidden state, get rid of the 'apply' button.  the
policy is what's in the textarea, and cajoling uses whatever is in
that textarea at the time.

I did (b) because that seems simpler.

2. I added a 'default' button in the policy tab, which reverts policy
to the default.

R=jasvir@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@4475 360991c0-9f3b-0410-ba87-8b362a00edf4