Playground needs to mark GWT's frame as feral http://codereview.appspot.com/4527099
So the future of FERAL_FRAME_OBJECT___ is under discussion, but in
the meantime, this change makes the Playground policy tab work again.

The reason it doesn't work is the policy is eval'ed in GWT's frame,
and that frame does not have the Object.FERAL_FRAME_OBJECT___ marker,
so anything defined in the policy will tame to undefined.

This change adds a call to initFeralFrame(), which will add that
marker.

R=jasvir@gmail.com, ihab.awad@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@4494 360991c0-9f3b-0410-ba87-8b362a00edf4