Fixes javascript URLs and inline event handlers http://codereview.appspot.com/4530099
Modifies caja.js to copy the taming frame's plugin_dispatchToHandler___
and plugin_dispatchEvent___ functions into the guest frame (which is
necessary for inline event handlers like onclick="...") and into
the window containing the dom elements (which is necessary for
'javascript:' URLs).  Since we always use a single taming frame
even when there are multiple guest frames, this fix should work in
the presence of multiple cajoled pages.

R=mikesamuel, jasvir


git-svn-id: https://google-caja.googlecode.com/svn/trunk@4499 360991c0-9f3b-0410-ba87-8b362a00edf4