fix [summarize] [Fatal Error] :-1:-1: Premature end of file. http://codereview.appspot.com/4634089
ant runtests sometimes succeeds, but fails at the end with:
[summarize] [Fatal Error] :-1:-1: Premature end of file.
[summarize] ERROR:  'Premature end of file.'

This happens when one of the TEST-*.xml files in ant-reports/tests
is empty, usually because the test was interrupted before writing
the report.  And for some reason the Xerces SAX parser throws an
error on files without any content.

So this change makes the summarize task ignore empty files.  You'll
still get the weird error if you somehow have a blank but nonempty
test results file, but that's much less likely to happen.

R=jasvir@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@4523 360991c0-9f3b-0410-ba87-8b362a00edf4